feat: don't coerce null values to undefined/empty in Joi validation
BREAKING CHANGE: ^^^
null values stay as null values now
They become INVALID in most of the places, unless explicitly allowed

Purpose of this change is to allow APIs to send null values to explicitly indicate
 "absense of value".

Example: a patch API that wants to "unset"/delete a value. It now can send { someProp: null } to
indicate that backend have to "delete" this proporty. Previously null value would be stripped and
 input become { } (empty object).